Refactor admin implementation and fix admin editting

+++ lessons/views.py
@@ -1,9 +1,11 @@
+from django.contrib.admin.options import PermissionDenied
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ObjectDoesNotExist
 from django.shortcuts import get_object_or_404, render, redirect
 from .models import Booking, Invoice, RequestForLessons, User, SchoolTerm, SchoolAdmin, User
 from .forms import (
+    EditAdminForm,
     RequestForLessonsForm,
     SchoolTermForm,
     StudentSignUpForm,
@@ -16,6 +18,7 @@ from .forms import (
     CreateAdminForm,
 )
 
+
 #  Create your views here.
 def home(request):
     if request.user.is_authenticated:
@@ -25,6 +28,9 @@ def home(request):
 
 
 def log_in(request):
+    if request.user.is_authenticated:
+        return redirect("account")
+
     if request.method == "POST":
         form = LogInForm(request.POST)
         if form.is_valid():
@@ -55,10 +61,16 @@ def forgot_password(request):
 
 
 def sign_up(request):
+    if request.user.is_authenticated:
+        return redirect("account")
+
     return render(request, "sign_up.html")
 
 
 def sign_up_student(request):
+    if request.user.is_authenticated:
+        return redirect("account")
+
     if request.method == "POST":
         form = StudentSignUpForm(request.POST)
         if form.is_valid():
@@ -71,20 +83,18 @@ def sign_up_student(request):
     return render(request, "sign_up_student.html", {"form": form})
 
 
-# check if the user is a director then display sign up page
-def sign_up_admin(request):
+def create_admin(request):
     if request.method == "POST":
-        form = CreateAdminForm(
-            request.POST
-        )  # creates a bound version of the form with post data
+        # create a bound version of the form with post data
+        form = CreateAdminForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect("account")
+            return redirect("admin_list")
     else:
         form = (
             CreateAdminForm()
         )  # create a form with CreateAdminForm constructor, pass that form to template to render it
-    return render(request, "sign_up_admin.html", {"form": form})
+    return render(request, "create_admin.html", {"form": form})
     # successful form means you save user record in database and redirect them to the database
 
 @login_required
@@ -107,15 +117,15 @@ def account(request):
     else:
         # UNRECOGNIZED USER TYPE
         # this shouldn't happen, log user out and send him to welcome page
-        # print("Who are you ?? " + str(request.user))
         logout(request)
-        return redirect("home")
+        return redirect("account")
 
 
 @login_required
 def bookings_list(request):
-    if request.user.is_student is False:
+    if not request.user.is_student:
         return redirect("account")
+
     bookings = request.user.student.booking_set.all()
     return render(request, "bookings_list.html", {"bookings": bookings})
 
@@ -133,26 +143,11 @@ def show_booking(request, booking_id):
 
 @login_required
 def requests_list(request):
-    if request.user.is_student is False:
-        return redirect("home")
+    if not request.user.is_student:
+        raise PermissionDenied
+
     requests = request.user.student.requestforlessons_set.all()
     return render(request, "requests_list.html", {"requests": requests})
-
-
-@login_required
-def view_admin_list(request):
-    if request.user.schooladmin.directorStatus:
-        admins = SchoolAdmin.objects.all()
-        return render(request, "admin_list.html", {"admins": admins})
-    elif request.user.schooladmin.deleteAdmins and request.user.schooladmin.editAdmins:
-        admins = SchoolAdmin.objects.all()
-        return render(request, "admin_list.html", {"admins": admins})
-    elif request.user.schooladmin.editAdmins:
-        admins = SchoolAdmin.objects.all()
-        return render(request, "admin_list_edit_only.html", {"admins": admins})
-    elif request.user.schooladmin.deleteAdmins:
-        admins = SchoolAdmin.objects.all()
-        return render(request, "admin_list_delete_only.html", {"admins": admins})
 
 
 @login_required
@@ -209,6 +204,9 @@ def fulfill_request(request, id):
 
 @login_required
 def create_request(request):
+    if not request.user.is_student:
+        raise PermissionDenied
+
     if request.method == "POST":
         form = RequestForLessonsForm(request.POST, student=request.user.student)
         if form.is_valid():
@@ -221,6 +219,9 @@ def create_request(request):
 
 @login_required
 def edit_request(request, id):
+    if not request.user.is_student:
+        raise PermissionDenied
+
     req = get_object_or_404(RequestForLessons, id=id)
     if request.method == "POST":
         form = RequestForLessonsForm(
@@ -236,80 +237,52 @@ def edit_request(request, id):
 
 
 @login_required
-def all_requests_list(request):
+def admin_list(request):
     if not request.user.is_school_admin:
-        return redirect("account")
+        raise PermissionDenied
 
-    all_requests = RequestForLessons.objects.all()
-    return render(request, "all_requests_list.html", {"all_requests": all_requests})
+    admins = SchoolAdmin.objects.all().exclude(user=request.user)
+    return render(
+        request, "admin_list.html", {"admins": admins, "current_user": request.user}
+    )
+
+    # if request.user.schooladmin.is_director:
+    #     return render(request, "admin_list.html", {"admins": admins})
+    # elif request.user.schooladmin.editAdmins:
+    #     return render(request, "admin_list_edit_only.html", {"admins": admins})
+    # elif request.user.schooladmin.deleteAdmins:
+    #     return render(request, "admin_list_delete_only.html", {"admins": admins})
+
 
 @login_required
-def all_bookings_list(request):
-    if not request.user.is_school_admin:
-        return redirect("account")
-
-    bookings = Booking.objects.all()
-    return render(request, "all_bookings_list.html", {"bookings": bookings})
-
-def admin_delete_booking(request, id):
-    req = get_object_or_404(Booking, id=id)
-    if req:
-        req.delete()
-        print("success!")
-
-    if request.user.is_school_admin:
-        return redirect("all_bookings_list")
-    else:
-        return redirect("account")
-
-
-def admin_edit_booking(request, id):
-    lesson_request = RequestForLessons.objects.get(id=id)
-
-    if request.method == "POST":
-        form = FulfillLessonRequestForm(request.POST, lesson_request=lesson_request)
-        if form.is_valid():
-            booking = form.save()
-            print(booking)
-            return redirect("all_bookings_list")
-    else:
-        form = FulfillLessonRequestForm(lesson_request=lesson_request)
-
-    student_name = (
-        lesson_request.student.user.first_name
-        + " "
-        + lesson_request.student.user.last_name
-    )
-
-    return render(
-        request,
-        "edit_booking_form.html",
-        {"request_id": id, "form": form, "student_name": student_name},
-    )
-
-
 def edit_admin(request, id):
-    currentadmin = get_object_or_404(User, pk=id)
+    if not (request.user.is_school_admin and request.user.schooladmin.is_director):
+        raise PermissionDenied
+
+    admin = get_object_or_404(SchoolAdmin, pk=id)
+
     if request.method == "POST":
-        form = CreateAdminForm(
-            request.POST, instance=currentadmin, schooladmin=request.user.schooladmin
-        )
+        form = EditAdminForm(request.POST, instance=admin)
         if form.is_valid():
-            currentadmin = form.save(edit=True)
-            currentadmin.save()
+            admin = form.save()
             return redirect("admin_list")
     else:
-        form = CreateAdminForm(instance=currentadmin)
-    return render(request, "edit_admin.html", {"admin_user_id": id, "form": form})
+        form = EditAdminForm(instance=admin)
+    return render(request, "edit_admin.html", {"admin_id": id, "form": form})
 
 
 @login_required
 def delete_admin(request, id):
-    currentadmin = SchoolAdmin.objects.get(pk=id)
-    if currentadmin:
-        currentadmin.user.is_active=False
-        currentadmin.delete()
-        return render(request, "delete_admin.html")
+    if not (request.user.is_school_admin and request.user.schooladmin.is_director):
+        raise PermissionDenied
+
+    admin = get_object_or_404(SchoolAdmin, pk=id)
+    if admin:
+        admin.delete()
+        print("success!")
+
+    return redirect("admin_list")
+
 
 
 def payment(request):
@@ -403,7 +376,8 @@ def select_child(request):
 @login_required
 def school_terms_list(request):
     if not request.user.is_school_admin:
-        return redirect("account")
+        raise PermissionDenied
+
     school_terms = SchoolTerm.objects.all()
     return render(request, "school_terms_list.html", {"school_terms": school_terms})
 
@@ -412,7 +386,7 @@ def school_terms_list(request):
 # TODO:  @adminrequired
 def create_school_term(request):
     if not request.user.is_school_admin:
-        return redirect("account")
+        raise PermissionDenied
 
     if request.method == "POST":
         form = SchoolTermForm(request.POST)
@@ -427,10 +401,14 @@ def create_school_term(request):
 
 @login_required
 def delete_school_term(request, id):
+    if not request.user.is_school_admin:
+        raise PermissionDenied
+
     term = SchoolTerm.objects.get(id=id)
     if term:
         term.delete()
-        print(f"Request {id} deleted")
+        print(f"SchoolTerm {id} deleted")
+
         return redirect("school_terms_list")
     print("cant find term")
 
@@ -438,7 +416,7 @@ def delete_school_term(request, id):
 @login_required
 def edit_school_term(request, id):
     if not request.user.is_school_admin:
-        return redirect("account")
+        raise PermissionDenied
 
     term = get_object_or_404(SchoolTerm, id=id)
